+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Improve:
 
 Remove:
 * remove: Remove 'tools4a-utils-ktx' module
+* remove: Activityx remove start, startByClass method
 
 Fix:
 * bug: Fix the bug that the Contextx wifiManagerOrNull() method did not convert the incoming Context to ApplicationContext